+++ fixtures/gists-app/app/components/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Meta, Scripts, Styles, Routes, useGlobalData } from "@remix-run/react";
 
 export default function App() {
-  let [data] = useGlobalData();
+  let data = useGlobalData();
 
   React.useEffect(() => {
     // We use this in the tests to wait for React to hydrate the page.